use size_t less
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5496 d7cf8633-e32d-0410-b094-e92efae38249

+++ snd_main.h
@@ -26,8 +26,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 typedef struct
 {
-	size_t	length;
-	size_t	offset;
+	int		length;
+	int		offset;
 	qbyte	data[4];	// variable sized
 } sfxbuffer_t;
 
@@ -59,7 +59,7 @@ struct sfx_s
 	unsigned int		flags;			// cf SFXFLAG_* defines
 	snd_format_t		format;
 	int					loopstart;
-	size_t				total_length;
+	int					total_length;
 	const snd_fetcher_t	*fetcher;
 	void				*fetcher_data;	// Per-sfx data for the sound fetching functions
 };
@@ -80,9 +80,9 @@ typedef struct
 	int				master_vol;		// 0-255 master volume
 	int				leftvol;		// 0-255 volume
 	int				rightvol;		// 0-255 volume
-	size_t			end;			// end time in global paintsamples
-	size_t			lastptime;		// last time this channel was painted
-	size_t			pos;			// sample position in sfx
+	int				end;			// end time in global paintsamples
+	int				lastptime;		// last time this channel was painted
+	int				pos;			// sample position in sfx
 	int				entnum;			// to allow overriding a specific sound
 	int				entchannel;
 	vec3_t			origin;			// origin of sound effect
